+++ llvm/lib/Target/AArch64/AArch64TargetMachine.cpp
@@ -422,6 +422,8 @@ AArch64TargetMachine::AArch64TargetMachine(const Target &T, const Triple &TT,
   // AArch64 supports the debug entry values.
   setSupportsDebugEntryValues(true);
 
+  setEnableSSAFastRegAlloc(true);
+
   // AArch64 supports fixing up the DWARF unwind information.
   if (!getMCAsmInfo().usesWindowsCFI())
     setCFIFixup(true);